Refactoring for browser cache v2/fetcher
