+++ fanficfare/configurable.py
@@ -40,9 +40,9 @@ except ImportError:
     chardet = None
 
 from . import exceptions
-from . import fetcher
-from . import nsapa_proxy
-from . import flaresolverr_proxy
+from . import fetchers
+from .fetchers import fetcher_nsapa_proxy
+from .fetchers import fetcher_flaresolverr_proxy
 
 ## has to be up here for brotli-dict to load correctly.
 from .browsercache import BrowserCache
@@ -592,7 +592,7 @@ class Configuration(ConfigParser):
         self.fetcher = None # the network layer for getting pages the
         self.sleeper = None
         # caching layer for getting pages, create one if not given.
-        self.basic_cache = basic_cache or fetcher.BasicCache()
+        self.basic_cache = basic_cache or fetchers.BasicCache()
         # don't create a browser cache by default.
         self.browser_cache = browser_cache
         self.filelist_fetcher = None # used for _filelist
@@ -999,7 +999,7 @@ class Configuration(ConfigParser):
             # always use base requests fetcher for _filelist--odds are
             # much higher user wants a file:// than something through
             # browser cache or a proxy.
-            self.filelist_fetcher = fetcher.RequestsFetcher(self.getConfig,
+            self.filelist_fetcher = fetchers.RequestsFetcher(self.getConfig,
                                                             self.getConfigList)
         ( data, redirecturl ) = self.filelist_fetcher.get_request_redirected(fn)
         retval = None
@@ -1029,19 +1029,19 @@ class Configuration(ConfigParser):
 
             if self.getConfig('use_flaresolverr_proxy',False):
                 logger.debug("use_flaresolverr_proxy:%s"%self.getConfig('use_flaresolverr_proxy'))
-                fetchcls = flaresolverr_proxy.FlareSolverr_ProxyFetcher
+                fetchcls = fetcher_flaresolverr_proxy.FlareSolverr_ProxyFetcher
                 if self.getConfig('use_flaresolverr_proxy') != 'withimages' and not self.getConfig('use_browser_cache'):
                     logger.warning("FlareSolverr v2+ doesn't work with images: include_images automatically set false")
                     logger.warning("Set use_flaresolverr_proxy:withimages if your are using FlareSolver v1 and want images")
                     self.set('overrides', 'include_images', 'false')
             elif self.getConfig('use_nsapa_proxy',False):
                 logger.debug("use_nsapa_proxy:%s"%self.getConfig('use_nsapa_proxy'))
-                fetchcls = nsapa_proxy.NSAPA_ProxyFetcher
+                fetchcls = fetcher_nsapa_proxy.NSAPA_ProxyFetcher
             elif self.getConfig('use_cloudscraper',False):
                 logger.debug("use_cloudscraper:%s"%self.getConfig('use_cloudscraper'))
-                fetchcls = fetcher.CloudScraperFetcher
+                fetchcls = fetchers.CloudScraperFetcher
             else:
-                fetchcls = fetcher.RequestsFetcher
+                fetchcls = fetchers.RequestsFetcher
             self.fetcher = fetchcls(self.getConfig,
                                     self.getConfigList)
 
@@ -1052,7 +1052,7 @@ class Configuration(ConfigParser):
 
             ## doesn't sleep when fromcache==True
             ## saved for set_sleep
-            self.sleeper = fetcher.SleepDecorator()
+            self.sleeper = fetchers.SleepDecorator()
             self.sleeper.decorate_fetcher(self.fetcher)
 
             ## cache decorator terminates the chain when found.
@@ -1065,17 +1065,17 @@ class Configuration(ConfigParser):
                     if self.browser_cache is None:
                         self.browser_cache = BrowserCache(self.getConfig("browser_cache_path"),
                                                           age_limit=self.getConfig("browser_cache_age_limit"))
-                    fetcher.BrowserCacheDecorator(self.browser_cache).decorate_fetcher(self.fetcher)
+                    fetchers.BrowserCacheDecorator(self.browser_cache).decorate_fetcher(self.fetcher)
                 except Exception as e:
                     logger.warning("Failed to setup BrowserCache(%s)"%e)
                     raise
             ## cache decorator terminates the chain when found.
             logger.debug("use_basic_cache:%s"%self.getConfig('use_basic_cache'))
             if self.getConfig('use_basic_cache') and self.basic_cache is not None:
-                fetcher.BasicCacheDecorator(self.basic_cache).decorate_fetcher(self.fetcher)
+                fetchers.BasicCacheDecorator(self.basic_cache).decorate_fetcher(self.fetcher)
 
             if self.getConfig('progressbar'):
-                fetcher.ProgressBarDecorator().decorate_fetcher(self.fetcher)
+                fetchers.ProgressBarDecorator().decorate_fetcher(self.fetcher)
         if cookiejar is not None:
             self.fetcher.set_cookiejar(cookiejar)
         return self.fetcher
